mcp-server-web/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
asyncpg==0.30.0
python-dotenv==1.1.1
pydantic==2.11.7
bcrypt==4.2.0
pyjwt==2.9.0
python-jose[cryptography]==3.3.0
email-validator==2.3.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
valkey==6.0.2
python-ulid==3.0.0
